adding to the ZombiesArena entropy, showing messages for personal bests

+++ zombies/src/main/java/io/github/zap/zombies/stats/map/MapStats.java
@@ -26,6 +26,6 @@ public class MapStats extends Stats<String> {
 
     }
 
-    Map<UUID, Integer> bestTimes = new HashMap<>();
+    Map<UUID, Long> bestTimes = new HashMap<>();
 
 }
